// src/app/Models/Employee.ts
export interface EmployeeM {
    id: number;
    name: string;
    email: string;
    phone: string;
    address: string;
    salary: number;

    // date: Date;
    // time: Date;
    // status: string;
    // image: string;
    // gender: string;
    // department: string;
    // position: string;
    // experience: string;
}